Add files via upload
In this project I want to create a visual representation of the rotation of the stars and the gravitational fluctuations.
I used the rotate function we learn in Matrix with for/if statements to draw the planets and modified the color and alpha. I added the fbo effect at the first scene to resemble to big bang. The central star is moving with the mouse, and the rest of the stars are following/escaping from the central star depends on the direction it moves toward. I tried to create the effect of gravity by using easing, but I failed to assign different speeds to different stars.